Almost complete redo of blockplacemods, small fixes
I've redone most of blockplacemods its code. It is now MUCH easier to
expand in the future and I think the code is more elegant and stuff. It
all works the same as before, except the command is now /toggle
<setting>
I've merged the setting for /autotakewater and /autofillcauldron into
one setting named cauldron.