Disabled vanishfix, preparing merge
I replaced /warn and /warnp commands with @simplecommand hook. they
happened to be missing the 'if' from the permission checks and this
takes care of some other things as well.

I added a module called vanish which will replace essentials vanish
which happens to be really buggy as you know.

I added a serversigns module but it is not added to main.py and thus it
will not cause problems!